refactor: split the SerializationService.java into JsonSchemaSerializationService and a serializer
