+++ digiwf-json-serialization-starter/src/main/java/io/muenchendigital/digiwf/json/serialization/configuration/JsonSerializationAutoConfiguration.java
@@ -1,6 +1,9 @@
 package io.muenchendigital.digiwf.json.serialization.configuration;
 
-import io.muenchendigital.digiwf.json.serialization.SerializationService;
+import io.muenchendigital.digiwf.json.serialization.JsonSchemaSerializationService;
+import io.muenchendigital.digiwf.json.serialization.serializer.JsonSchemaBaseSerializer;
+import io.muenchendigital.digiwf.json.serialization.serializer.JsonSchemaSerializer;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -9,8 +12,14 @@ import org.springframework.context.annotation.Configuration;
 public class JsonSerializationAutoConfiguration {
 
     @Bean
-    public SerializationService serializationService() {
-        return new SerializationService();
+    @ConditionalOnMissingBean
+    public JsonSchemaBaseSerializer jsonSchemaSerializer() {
+        return new JsonSchemaSerializer();
+    }
+
+    @Bean
+    public JsonSchemaSerializationService jsonSchemaSerializationService(final JsonSchemaBaseSerializer serializer) {
+        return new JsonSchemaSerializationService(serializer);
     }
 
 }
